Add documentation to color palettes and split out vector operations.
